Factor WillPopScope out of Form
This functionality is independently useful.

Fixes #8293
